added a singleEvent reducer
