+++ src/components/single-event.js
@@ -27,6 +27,7 @@ class SingleEvent extends React.Component {
     rsvpCheck() {
         // arr is an array of objects with the user id and username of all rsvp'd users
         let arr = this.props.attending;
+        console.log(arr);
         if (arr === null) {
             // if the array is null, no one has rsvp'd
             return false
@@ -117,8 +118,8 @@ class SingleEvent extends React.Component {
 
 const mapStateToProps = state => {
     return {
-        attending: state.event.attending,
-        event: state.event.selectedEvent,
+        attending: state.singleEvent.attending,
+        event: state.singleEvent.selectedEvent,
         username: state.auth.currentUser.username
     };
 };
